[ADD]When the Agent version is inconsistent, limit the deployment of applications and create networks, domain names

+++ src/main/java/io/choerodon/devops/app/service/impl/DevopsIngressServiceImpl.java
@@ -30,6 +30,8 @@ import io.choerodon.devops.infra.common.util.enums.ObjectType;
 import io.choerodon.devops.infra.dataobject.DevopsIngressDO;
 import io.choerodon.devops.infra.dataobject.DevopsIngressPathDO;
 import io.choerodon.mybatis.pagehelper.domain.PageRequest;
+import io.choerodon.websocket.helper.EnvListener;
+import io.choerodon.websocket.helper.EnvSession;
 
 
 /**
@@ -41,6 +43,7 @@ import io.choerodon.mybatis.pagehelper.domain.PageRequest;
 @Component
 public class DevopsIngressServiceImpl implements DevopsIngressService {
     private static final String PATH_ERROR = "error.path.empty";
+    private static final String ENV_DISCONNECTED = "error.env.disconnect";
 
     private static JSON json = new JSON();
     private IDevopsIngressService idevopsIngressService;
@@ -48,6 +51,7 @@ public class DevopsIngressServiceImpl implements DevopsIngressService {
     private DevopsServiceRepository devopsServiceRepository;
     private DevopsEnvironmentRepository environmentRepository;
     private DevopsEnvCommandRepository devopsEnvCommandRepository;
+    private EnvListener envListener;
 
     /**
      * 构造函数
@@ -56,95 +60,104 @@ public class DevopsIngressServiceImpl implements DevopsIngressService {
                                     DevopsServiceRepository devopsServiceRepository,
                                     DevopsEnvironmentRepository devopsEnvironmentRepository,
                                     IDevopsIngressService idevopsIngressService,
-                                    DevopsEnvCommandRepository devopsEnvCommandRepository) {
+                                    DevopsEnvCommandRepository devopsEnvCommandRepository,
+                                    EnvListener envListener) {
         this.devopsIngressRepository = devopsIngressRepository;
         this.devopsServiceRepository = devopsServiceRepository;
         this.environmentRepository = devopsEnvironmentRepository;
         this.idevopsIngressService = idevopsIngressService;
         this.devopsEnvCommandRepository = devopsEnvCommandRepository;
+        this.envListener = envListener;
     }
 
     @Override
     public void addIngress(DevopsIngressDTO devopsIngressDTO, Long projectId) {
-        DevopsIngressValidator.checkAppVersion(devopsIngressDTO);
-        String name = devopsIngressDTO.getName();
-        String domain = devopsIngressDTO.getDomain();
-        Long envId = devopsIngressDTO.getEnvId();
-        DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(envId);
-        DevopsIngressDO devopsIngressDO = new DevopsIngressDO(projectId, envId, domain, name);
-        List<DevopsIngressPathDO> devopsIngressPathDOS = new ArrayList<>();
-        V1beta1Ingress ingress = createIngress(domain, name, devopsEnvironmentE.getCode());
-        if (devopsIngressDTO.getDevopsIngressPathDTOList().isEmpty()) {
-            throw new CommonException(PATH_ERROR);
-        }
-        devopsIngressDTO.getDevopsIngressPathDTOList().forEach(t -> {
-            if (t.getPath() == null || t.getServiceId() == null) {
-                throw new CommonException(PATH_ERROR);
-            }
-
-            DevopsServiceE devopsServiceE = getDevopsService(t.getServiceId());
-
-            devopsIngressPathDOS.add(new DevopsIngressPathDO(
-                    devopsIngressDTO.getId(), t.getPath(), devopsServiceE.getId(), devopsServiceE.getName()));
-            ingress.getSpec().getRules().get(0).getHttp().addPathsItem(createPath(t.getPath(), t.getServiceId()));
-        });
-        DevopsEnvCommandE devopsEnvCommandE = DevopsEnvCommandFactory.createDevopsEnvCommandE();
-        devopsEnvCommandE.setObject(ObjectType.INGRESS.getObjectType());
-        devopsEnvCommandE.setObjectId(devopsIngressDO.getId());
-        devopsEnvCommandE.setCommandType(CommandType.CREATE.getCommandType());
-        devopsEnvCommandE.setStatus(CommandStatus.DOING.getCommandStatus());
-        devopsIngressRepository.createIngress(devopsIngressDO, devopsIngressPathDOS);
-        idevopsIngressService.createIngress(json.serialize(ingress),
-                name,
-                devopsEnvironmentE.getCode(),
-                devopsEnvCommandRepository.create(devopsEnvCommandE).getId());
-
-
-    }
-
-    @Override
-    public void updateIngress(Long id, DevopsIngressDTO devopsIngressDTO, Long projectId) {
-        DevopsIngressValidator.checkAppVersion(devopsIngressDTO);
-        DevopsIngressDTO ingressDTO = devopsIngressRepository.getIngress(projectId, id);
-        if (!devopsIngressDTO.equals(ingressDTO)) {
+        if (isEnvConnected(devopsIngressDTO.getEnvId())) {
+            DevopsIngressValidator.checkAppVersion(devopsIngressDTO);
             String name = devopsIngressDTO.getName();
             String domain = devopsIngressDTO.getDomain();
-            Long domainEnvId = devopsIngressDTO.getEnvId();
-            DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(domainEnvId);
-
+            Long envId = devopsIngressDTO.getEnvId();
+            DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(envId);
+            DevopsIngressDO devopsIngressDO = new DevopsIngressDO(projectId, envId, domain, name);
+            List<DevopsIngressPathDO> devopsIngressPathDOS = new ArrayList<>();
+            V1beta1Ingress ingress = createIngress(domain, name, devopsEnvironmentE.getCode());
             if (devopsIngressDTO.getDevopsIngressPathDTOList().isEmpty()) {
                 throw new CommonException(PATH_ERROR);
             }
-            V1beta1Ingress ingress = createIngress(domain, name, devopsEnvironmentE.getCode());
-            DevopsIngressDO devopsIngressDO = new DevopsIngressDO(
-                    id, projectId, domainEnvId, domain, name);
-            List<DevopsIngressPathDO> devopsIngressPathDOS = new ArrayList<>();
             devopsIngressDTO.getDevopsIngressPathDTOList().forEach(t -> {
-                if (t.getPath() == null) {
+                if (t.getPath() == null || t.getServiceId() == null) {
                     throw new CommonException(PATH_ERROR);
-                } else if (t.getServiceId() == null) {
-                    throw new CommonException("error.service.id.get");
                 }
 
                 DevopsServiceE devopsServiceE = getDevopsService(t.getServiceId());
 
                 devopsIngressPathDOS.add(new DevopsIngressPathDO(
-                        id, t.getPath(), devopsServiceE.getId(), devopsServiceE.getName()));
-                ingress.getSpec().getRules().get(0).getHttp()
-                        .addPathsItem(createPath(t.getPath(), t.getServiceId()));
+                        devopsIngressDTO.getId(), t.getPath(), devopsServiceE.getId(), devopsServiceE.getName()));
+                ingress.getSpec().getRules().get(0).getHttp().addPathsItem(createPath(t.getPath(), t.getServiceId()));
             });
-            DevopsEnvCommandE devopsEnvCommandE = devopsEnvCommandRepository
-                    .queryByObject(ObjectType.INGRESS.getObjectType(), id);
-            devopsEnvCommandE.setCommandType(CommandType.UPDATE.getCommandType());
+            DevopsEnvCommandE devopsEnvCommandE = DevopsEnvCommandFactory.createDevopsEnvCommandE();
+            devopsEnvCommandE.setObject(ObjectType.INGRESS.getObjectType());
+            devopsEnvCommandE.setObjectId(devopsIngressDO.getId());
+            devopsEnvCommandE.setCommandType(CommandType.CREATE.getCommandType());
             devopsEnvCommandE.setStatus(CommandStatus.DOING.getCommandStatus());
-            devopsEnvCommandRepository.update(devopsEnvCommandE);
-            if (!ingressDTO.getName().equals(name)) {
-                idevopsIngressService.deleteIngress(ingressDTO.getName(), devopsEnvironmentE.getCode(),devopsEnvCommandE.getId());
-            }
-            devopsIngressRepository.updateIngress(devopsIngressDO, devopsIngressPathDOS);
+            devopsIngressRepository.createIngress(devopsIngressDO, devopsIngressPathDOS);
             idevopsIngressService.createIngress(json.serialize(ingress),
-                    name, devopsEnvironmentE.getCode(),devopsEnvCommandE.getId());
+                    name,
+                    devopsEnvironmentE.getCode(),
+                    devopsEnvCommandRepository.create(devopsEnvCommandE).getId());
+        } else {
+            throw new CommonException(ENV_DISCONNECTED);
+        }
 
+    }
+
+    @Override
+    public void updateIngress(Long id, DevopsIngressDTO devopsIngressDTO, Long projectId) {
+        if (isEnvConnected(devopsIngressDTO.getEnvId())) {
+            DevopsIngressValidator.checkAppVersion(devopsIngressDTO);
+            DevopsIngressDTO ingressDTO = devopsIngressRepository.getIngress(projectId, id);
+            if (!devopsIngressDTO.equals(ingressDTO)) {
+                String name = devopsIngressDTO.getName();
+                String domain = devopsIngressDTO.getDomain();
+                Long domainEnvId = devopsIngressDTO.getEnvId();
+                DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(domainEnvId);
+
+                if (devopsIngressDTO.getDevopsIngressPathDTOList().isEmpty()) {
+                    throw new CommonException(PATH_ERROR);
+                }
+                V1beta1Ingress ingress = createIngress(domain, name, devopsEnvironmentE.getCode());
+                DevopsIngressDO devopsIngressDO = new DevopsIngressDO(
+                        id, projectId, domainEnvId, domain, name);
+                List<DevopsIngressPathDO> devopsIngressPathDOS = new ArrayList<>();
+                devopsIngressDTO.getDevopsIngressPathDTOList().forEach(t -> {
+                    if (t.getPath() == null) {
+                        throw new CommonException(PATH_ERROR);
+                    } else if (t.getServiceId() == null) {
+                        throw new CommonException("error.service.id.get");
+                    }
+
+                    DevopsServiceE devopsServiceE = getDevopsService(t.getServiceId());
+
+                    devopsIngressPathDOS.add(new DevopsIngressPathDO(
+                            id, t.getPath(), devopsServiceE.getId(), devopsServiceE.getName()));
+                    ingress.getSpec().getRules().get(0).getHttp()
+                            .addPathsItem(createPath(t.getPath(), t.getServiceId()));
+                });
+                DevopsEnvCommandE devopsEnvCommandE = devopsEnvCommandRepository
+                        .queryByObject(ObjectType.INGRESS.getObjectType(), id);
+                devopsEnvCommandE.setCommandType(CommandType.UPDATE.getCommandType());
+                devopsEnvCommandE.setStatus(CommandStatus.DOING.getCommandStatus());
+                devopsEnvCommandRepository.update(devopsEnvCommandE);
+                if (!ingressDTO.getName().equals(name)) {
+                    idevopsIngressService.deleteIngress(ingressDTO.getName(), devopsEnvironmentE.getCode(), devopsEnvCommandE.getId());
+                }
+                devopsIngressRepository.updateIngress(devopsIngressDO, devopsIngressPathDOS);
+                idevopsIngressService.createIngress(json.serialize(ingress),
+                        name, devopsEnvironmentE.getCode(), devopsEnvCommandE.getId());
+
+            }
+        } else {
+            throw new CommonException(ENV_DISCONNECTED);
         }
     }
 
@@ -160,14 +173,18 @@ public class DevopsIngressServiceImpl implements DevopsIngressService {
 
     @Override
     public void deleteIngress(Long ingressId) {
-        DevopsEnvCommandE devopsEnvCommandE = devopsEnvCommandRepository
-                .queryByObject(ObjectType.INGRESS.getObjectType(), ingressId);
-        devopsEnvCommandE.setCommandType(CommandType.DELETE.getCommandType());
-        devopsEnvCommandE.setStatus(CommandStatus.DOING.getCommandStatus());
-        devopsEnvCommandRepository.update(devopsEnvCommandE);
-        DevopsIngressDO ingressDO = devopsIngressRepository.getIngress(ingressId);
-        DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(ingressDO.getEnvId());
-        idevopsIngressService.deleteIngress(ingressDO.getName(), devopsEnvironmentE.getCode(), devopsEnvCommandE.getId());
+        if (isEnvConnected(devopsIngressRepository.getIngress(ingressId).getEnvId())) {
+            DevopsEnvCommandE devopsEnvCommandE = devopsEnvCommandRepository
+                    .queryByObject(ObjectType.INGRESS.getObjectType(), ingressId);
+            devopsEnvCommandE.setCommandType(CommandType.DELETE.getCommandType());
+            devopsEnvCommandE.setStatus(CommandStatus.DOING.getCommandStatus());
+            devopsEnvCommandRepository.update(devopsEnvCommandE);
+            DevopsIngressDO ingressDO = devopsIngressRepository.getIngress(ingressId);
+            DevopsEnvironmentE devopsEnvironmentE = environmentRepository.queryById(ingressDO.getEnvId());
+            idevopsIngressService.deleteIngress(ingressDO.getName(), devopsEnvironmentE.getCode(), devopsEnvCommandE.getId());
+        } else {
+            throw new CommonException(ENV_DISCONNECTED);
+        }
 
     }
 
@@ -231,5 +248,16 @@ public class DevopsIngressServiceImpl implements DevopsIngressService {
             throw new CommonException("error.service.select");
         }
         return devopsServiceE;
+    }
+
+
+    Boolean isEnvConnected(Long envId) {
+        Map<String, EnvSession> envs = envListener.connectedEnv();
+        List<Long> envIds = new ArrayList<>();
+        for (Map.Entry<String, EnvSession> entry : envs.entrySet()) {
+            EnvSession envSession = entry.getValue();
+            envIds.add(envSession.getEnvId());
+        }
+        return envIds.contains(envId);
     }
 }
